Update FT config and policy for Fallback applyOn/skipOn
* Added FallbackConfig21 to understand applyOn/skipOn
* Added FallbackConfig to the annotation config factory
* Updated Interceptor and Extension to use factory to create FallbackConfig
* Updated FallbackPolicy to hold applyOn/skipOn values
