+++ dev/com.ibm.ws.microprofile.faulttolerance.spi/src/com/ibm/ws/microprofile/faulttolerance/spi/FallbackPolicy.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017 IBM Corporation and others.
+ * Copyright (c) 2017, 2019 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -24,4 +24,25 @@ public interface FallbackPolicy {
     public void setFallbackHandler(Class<? extends FallbackHandler<?>> clazz, FallbackHandlerFactory factory);
 
     public FallbackHandlerFactory getFallbackHandlerFactory();
+
+    /**
+     * Define the apply criteria
+     *
+     * @return the apply exception
+     */
+    public Class<? extends Throwable>[] getApplyOn();
+
+    @SuppressWarnings("unchecked")
+    public void setApplyOn(Class<? extends Throwable>... failOn);
+
+    /**
+     * Define the skip criteria
+     *
+     * @return the skip exception
+     */
+    public Class<? extends Throwable>[] getSkipOn();
+
+    @SuppressWarnings("unchecked")
+    public void setSkipOn(Class<? extends Throwable>... skipOn);
+
 }
